Add make_archive
This commit adds `make_archive` functionality to chainerio.
This function shares the same interfaces and works similar to
the standard function `shutil.make_archive`.
In addition, this function can also make archive to the hdfs and other
filesystem, by calling the `shutil.make_archive` to make the archive to a
local tmp file and then copy the archive to the final destination on
hdfs or other filesystem.

Limitation:
The base directory has to be on the posix compliant filesystem for
`shutil.make_archive` to access.
This can be solved by copying the files to local before making the
archive.