Fix bugs in imosci script
This commit fixes a bug in imosci script.
In the previous CI, there are two tests performed, tox for pytest and
sphinx build. The CI script fails to combine all the return value
of the test commands, hence the CI status only show the status of the
sphinx build.

In this commit, the return status of both tests are combined and returned.