Add nested zip support for Python < 3.7
This commit adds nested zip support for Python< 3.7.
Unlike the Python > 3.7, due to implementation issues in Python,
the nested zip support is not enabled by default.
A workaround is added to support that, which possible performance and
memory consumption issues.
A warning message is generated in that case.
For more details, please refer to the comments in
`containers/zip.py`.