Bugfix: list and isdir for non-directory-included zip
This PR fixes a bug which is related to non-directory-included zip, a
kind of zip that contains hierarchical structure but does not contain
the directories themselves. Such kind of zip can be made with "zip -D".

This PR fixes the bug when list or isdir to these "non-exist" directory
in such non-directory-included zip.